uber-page: Fix a bug where sections can show up unexpectedly when a search-result is active.
When a section is hidden by default (e.g. bluetooth), it is not marked
'search-hidden' after a search. So when the system marks it as visible (e.g.
BluetoothOptionsHandler detects bluetooth is available), it is made visible even
if it doesn't match the search criteria. So make sure 'search-hidden' is set
even for the sections that are hidden by default.

BUG=126201
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10365008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@135725 0039d316-1c4b-4281-b951-d872f2087c98