Don't make sad tab view for normal renderer termination
We noticed some of the statistics for Tabs.SadTab.CrashCreated were much higher than expected on R19.  Something changed that caused us to create a SadTabView every time a tab was closed, even if the renderer exited normally.  This patch skips the creation of the SadTabView for normal termination status.

BUG=126507
TEST=manual, opened and closed tabs manually and verified about:histograms Tabs.SadTab.CrashCreated does not go up, also tested killing tabs with task manager and kill -11 to the renderer process to verify that sad tabs still work in those cases

Review URL: https://chromiumcodereview.appspot.com/10388062

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@136310 0039d316-1c4b-4281-b951-d872f2087c98