Simplify SplitStringIntoParagraphs().
The new code simply splits on all paragraph separators instead of keeping some strings containing runs of newlines.

This new approach removes the need for the code to workaround http://crbug.com/126297, so that when the CL that fixes that bug lands, it will not break this code.

BUG=126297
TEST=No visual change in JS alerts.

Review URL: https://chromiumcodereview.appspot.com/10383170

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137029 0039d316-1c4b-4281-b951-d872f2087c98