Fix blank lines being ignored in JS alerts.
I goofed when testing my change in http://crrev.com/137029, which had broken this.

The issue is that _multiline_ labels with the empty string as the text actually do have a height of 0. (Only non-multiline labels have a height of one line even if their content is the empty string.)

This change fixes the issue by not setting the multiline flag for empty lines.

BUG=126297
TEST=Try an alert() with \n's in it.

Review URL: https://chromiumcodereview.appspot.com/10310176

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137307 0039d316-1c4b-4281-b951-d872f2087c98