Fix output locations of syzygy versions of mini_installer's PDB and Map files.
The default paths have changed under from Visual Studio 2008 to 2010 and which causes the PDB and MAP files for the syzygyized mini_installer to get generated to the same locations as for the regular mini_installer. This further causes the staging of the Syzygy optimized build to fail because it expects the PDB file to be somewhere else.

Made the paths of the PDB and MAP files explicit.


BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10407006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137580 0039d316-1c4b-4281-b951-d872f2087c98