Add common code for extensions event filtering.
FilteredEventRouter will be used in the browser and the renderer to match
events to interested parties. In the browser it will be used to send events
only to renderers that are interested in them, and in the renderer it will be
used to send events to the correct JS listeners.


BUG=121479


Review URL: https://chromiumcodereview.appspot.com/10388135

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137829 0039d316-1c4b-4281-b951-d872f2087c98