Omnibox: search provider: fix timing histogram
The current implementation of the timing histogram causes a DCHECK when search provider suggest requests switch from success to failure or vice versa.  This fixes the problem.

I cannot test this to see if it works.  I think it should.  (The only machine I have configured to build chrome on is remote, so I cannot play with whether its network is enabled or not lest I lose my connection to it permanently.)

BUG=128613
TEST=


Review URL: https://chromiumcodereview.appspot.com/10382224

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137943 0039d316-1c4b-4281-b951-d872f2087c98