Base: Add power requirements to the System monitor.
The only thing we need in the short term is CPU_REQUIRED,
and that part is implemented for all platforms (as far as it
makes sense). The other options are implemented for Windows,
but given that currently there's no user, I expect the
implementation for other platforms to be written when there's
an actual need for it.

BUG=126591
TEST=none
Review URL: https://chromiumcodereview.appspot.com/10406012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137960 0039d316-1c4b-4281-b951-d872f2087c98