Implement Mac pref-pane UI mockups
This updates the screens to match the mockups.  There are two
new nib files.  One is a view with just a disable button.  The
other is a view with the PIN-confirmation controls.  The main
nib has been updated to remove all controls except for the
Status text and the padlock.  A new placeholder box has been
added in the middle of the window, which can show either of
the new nibs, depending on whether a new config has been
received by the pref-pane.

This also fixes some small issues:

PIN is reset whenever a new config is sent to the pref-pane -
this prevents the previous PIN dots being shown if the user
changes PIN a second time without closing the System Preferences
applet.

The PIN entry box is disabled until the pane is unlocked, for a
slightly improved UX.

BUG=128813,129032
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/10411032

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@138322 0039d316-1c4b-4281-b951-d872f2087c98