Makes the tabstrip switch between stacked and shrink modes based on whether the a mouse or touch input is used. Heres the specific semantics used: . If we detect 3 mouse moves within 200ms we assume the user is using   a mouse and switch to shrink. . On release we switch immediately based on the flags in the event.
Ideally we would switch to stacked on mouse enter, but because windows
generates events without telling us they are from a touch device we
have to infer based on time. When we convert all views to touch events
we can likely simplify this code.

BUG=128741
TEST=none
R=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10413067

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@138608 0039d316-1c4b-4281-b951-d872f2087c98