Add missing initialisation in GaiaOAuthClient::Core.
The changes to address crbug.com/128102 contained a bug whereby the caller-supplied OAuth URLs were not stored to the Core, leading to empty URLs being passed to the URLFetcher.  Confusingly, this provoked a URLFetcher response code of net::IO_PENDING, so this CL also adds a CHECK() to remoting/host/url_fetcher.cc to ensure that the supplied URL is valid.

BUG=129449

Review URL: https://chromiumcodereview.appspot.com/10431005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@138611 0039d316-1c4b-4281-b951-d872f2087c98