Implement GetContentsView in WebDialogView.
Implementation of GetContensView was deleted in http://src.chromium.org/viewvc/chrome?view=rev&revision=138631,
but it's necessary for Widget that does not require non-client view (e.g. TYPE_WINDOW_FRAMELESS) to be correctly initialized with WebView.
See the implementation of Widget::Init for more details.

This fixes the consequent crash that occurs when trying to show the keyboard overlay.

BUG=129662
TEST=Checked pressing Ctrl+Alt+/ did not cuase a crash

Review URL: https://chromiumcodereview.appspot.com/10442026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@139062 0039d316-1c4b-4281-b951-d872f2087c98