[Mac] Fix DCHECK(![tab inRapidClosureMode]) firing in tab_strip_drag_controller.mm.
I originally added that DCHECK in http://crrev.com/101209 out of concern that the call on the next line would trigger the -inRapidClosureMode code in -[TabView mouseUp:], which could close the tab leading to badness in the -endDrag: code.

Since the only reason this code calls -[TabView mouseUp:] is to ensure the tab gets selected, this change does that directly instead of going through -mouseUp:. This removes the need for that DCHECK since the original concern no longer applies.

BUG=114627
TEST=No DCHECK() should fire when following steps listed in http://crbug.com/114627.

Review URL: https://chromiumcodereview.appspot.com/10447058

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@139352 0039d316-1c4b-4281-b951-d872f2087c98