Remove NPChannelBase from cache on error.
Normal channel shutdown (as for navigation) goes via RemoveRoute(),
which removes the channel from the cache.  It looks like
OnChannelError() is expected to call RemoveRoute() at some point
during cleanup, but sometimes the channel is left in the cache and
re-vended (as measured by previous instrumentation).  Defensively
rename channels which see an error so that they aren't found in the
cache in the future.

BUG=97285
TEST=no user-visible change


Review URL: https://chromiumcodereview.appspot.com/10544075

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@141279 0039d316-1c4b-4281-b951-d872f2087c98