Remove the usage of TabRenderWatcher in WebDialogView.
The aim of the move is to remove the dependency of WebDialogView on src/chrome so that it can eventually go out of src/chrome.
WebDialogView does nothing in the callback functions and no class overrides them, so it's unnecessary for WebDialogView to use TanRenderWatcher.

BUG=124222
TEST=None

Review URL: https://chromiumcodereview.appspot.com/10502008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@142263 0039d316-1c4b-4281-b951-d872f2087c98