gdata: Convert GDataFileSystem::AddUploadFile() to asynchronous.
This is part 3 of converting public synchronous functions in GDataFileSystem to asynchronous.

Convert GDataFileSystem::AddUploadFile() to an asynchronous function. Upon completion it will invoke the callback on the UI thread.

Pass scoped_ptr<UploadFileInfo> instead of a UploadFileInfo* pointer to GDataUploader::UploadFailed, UploadCompletedCallback callbacks, and to the callback function for GDataFileSystem::AddUploadedFile, to signify transfer of ownership.

Decouple the UploadFileInfo entry removal in UploadFileInfoMap |pending_uploads_| from the deletion of the UploadFileInfo object.

Rename GDataUploader::DeleteUpload() to GDataUploader::RemoveUpload() and make it private. This function only removes entry from |pending_uploads_| but no longer deletes the UploadFileInfo object.

Defer deletion of UploadFileInfo object until the asynchronous function call is complete. Object is automatically deleted when the scoped pointer is no longer passed on.

BUG=132389
TEST=unit_tests, browser_tests

Review URL: https://chromiumcodereview.appspot.com/10540132

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@142515 0039d316-1c4b-4281-b951-d872f2087c98