Reduce GpuMemoryManager allocation spam.
The GpuMemoryManager will periodically schedule a Manage which calculates and assigns memory allocation limits for client contexts.  At the moment, this process is kicked off when tab visibility changes, which is relatively infrequent.

However, the future direction is for the GpuMemoryManager to track and manage client allocations as memory usage changes,  which may happen with each generated frame, once for each client.  We would like to be able to limit the rate at which we assign allocations, which for now is set to once every 16ms.

BUG=120246
TEST=Manual


Review URL: https://chromiumcodereview.appspot.com/10535063

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@142972 0039d316-1c4b-4281-b951-d872f2087c98