Add core plumbing for Instant Extended work
Adds UI model, types, and controller logic.  The controller logic is
encoded in search_tab_helper.cc/h.  The model observer changes are
multiplexed from the active tab through the browser's model via
search_delegate.h/cc.

The primary state in the model is the "mode" which changes depending on
interactions with the Omnibox and active URL of the page.  This state is
communicated from the tab to various "Top Chrome" components such as the
toolbar, the tab strip, etc.

BUG=133506
TEST=GoogleUtilTest.IsEmbeddedGoogleSearchUrl, SearchDelegateTest.*
R=sky@chromium.org, ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10644002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@143638 0039d316-1c4b-4281-b951-d872f2087c98