Demote LOG(WARNING) to VLOG(1) for an error message that is rather informational.
The error message is shown on almost every chromeos boot because this policy
is set to something only very rarely. Therefore this is not really an error
state but rather normal operational state.

Drive by: Cleaned up the rest of the VLOGs spread around the file.

BUG=112770
TEST=Less log abuse.


Review URL: https://chromiumcodereview.appspot.com/10666024

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@143906 0039d316-1c4b-4281-b951-d872f2087c98