sync: Move invariant check logic into Directory
The goal of this change is to allow runtime changes to the level of
invariant checking.  This will allow us to write unit tests which better
protect us against issue 125381.  We failed to prevent issue 134934 in
part because our tests against this scenario were not sufficiently
strict.  In the future, it may allow us to raise the level of checking
that we enforce in test code by letting us increase the strictness on a
case-by-case basis, thus saving us the trouble of having to fix all our
tests at once.

Along the way, this change modifies the way the invariant-checking
functions are invoked.  The end result is a bit simpler, though it does
have the side-effect of slightly increasing the strictness of the tests.
The CHANGES_ONLY checks will now ensure that all modified entries have
valid parents. This level of checking would previously be performed only
if the item's parent was also changed.

The change also removes the invariant check timeout.

BUG=134934
TEST=sync_unit_tests, unit_tests --gtest_filter='*Sync*'

Review URL: https://chromiumcodereview.appspot.com/10693041

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@145032 0039d316-1c4b-4281-b951-d872f2087c98