Shifted DownloadFile to exporting DownloadInterruptReasons.
DownloadFile used to return net::Errors (as BaseFile does).  With the 
folding of functionality from the DownloadFileManager implied by 
the ByteStream refactor (the DownloadFile gives completion messages
directly to the DownloadManager), it exported some DownloadInterruptReasons.
This CL takes it the rest of the way.

BUG=123998



Review URL: https://chromiumcodereview.appspot.com/10696043

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@145376 0039d316-1c4b-4281-b951-d872f2087c98