PPAPI/NaCl: Make ipc_channel_nacl post to the correct thread.
Connect is called on the thread where Channel::Send will be invoked (and where messages should be dispatched). But Channel might be _created_ on some other thread. Therefore, we need to post data that we read to the thread on which Connect is called. NOT the thread where the Channel was created.

This fixes an issue where sync messages were never getting their replies, because the DidRecvMsg task was being run on the thread where ChannelProxy::Send was called, not the Channel thread (which is the IO thread in this case).

BUG=116317
TEST=


Review URL: https://chromiumcodereview.appspot.com/10698138

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@146006 0039d316-1c4b-4281-b951-d872f2087c98