Change style of ConstrainedWindowViews. Removes the "baby blue" bar and close button in certain dialogs.  Instead, these dialogs will draw with no frame and the dialogs themselves will now be responsible for providing the ability to close.
Impact on derived classes:
Print preview has a cancel button already, but we'll add a close button in a separate CL.
This CL adds a close button in the web intents picker.
The htaccess dialog already has a cancel button and is frameless on Mac and Linux without a close button.  I don't think a change is required.


BUG=133363
TEST=Verify 133363

Review URL: https://chromiumcodereview.appspot.com/10631010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@146466 0039d316-1c4b-4281-b951-d872f2087c98