Make more room for theme_resources.grd resource IDs
Partly because of a grit bug that skips resources when there are <if> statements,
merging the grd files left us with a range of >600 resource ids in theme_resources.
This caused only sporadic failures, maybe because of the gaps in the IDs. Example:

http://build.chromium.org/p/tryserver.chromium/builders/linux_chromeos/builds/26848/steps/unit_tests/logs/stdio

ui_resources is okay for now with < 400 IDs allocated.


BUG=None
TEST=None


Review URL: https://chromiumcodereview.appspot.com/10789017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@146861 0039d316-1c4b-4281-b951-d872f2087c98