Resubmitting 10752006 - Adding field trial to enable speculative resource prefetching learning.
The CL was reverted because it broke the browser tests on chromeos.
This happened due to a couple of bugs occurring simultaneously.

1. The test faile because ResourcePrefetchPredictorObserver destructor had a
thread check that was wrong. This is fixed by 10779002.

2. The default group number in prerender_field_trial was initialized to -1
which clashes with the constant base::FieldTrial::kNotFinalized. This lead to
the experiment being enabled in the tests when it should not have been. That
is fixed in this CL.


BUG=None
TEST=None


Review URL: https://chromiumcodereview.appspot.com/10782006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@147091 0039d316-1c4b-4281-b951-d872f2087c98