Add rudimentary support in the ChromeFrame ActiveX to enable it work in a generic client site, i.e. a non IE client site. Some of our users are creating apps which host ChromeFrame to render web pages. This enables them to use the chrome network stack by setting up a privileged chrome frame container and setting the corresponding properties on the ChromeFrame ActiveX.
It mostly works well. However things like window.open break down as we need the ability in the ActiveX
to be able to connect to an existing ExternalTabContainer instance. This works well in IE where the IWebBrowser
implementation initiates the navigation correctly. However a generic implementation of the IWebBrowser interface
does not do this.

Fixes are to add a new function to the IChromeFramePrivileged interface which allows hosts to pass in the
navigation url early in the lifetime of the ActiveX. In the ChromeFrame ActiveX we check if the url coming in
is for an attach tab request and if so initiate the same.

Fixes portions of bug https://code.google.com/p/chromium/issues/detail?id=137335

BUG=137335
Review URL: https://chromiumcodereview.appspot.com/10783007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@147109 0039d316-1c4b-4281-b951-d872f2087c98