cros: Fix FramePainterTest.Basics failure on VS2010
The test was asserting that FramePainter::instances_ was NULL, but that may not be true if another test has run that created a FramePainter.  Perhaps VS2010 has changed the order in which this test runs.  Removed that assertion and reworked the test slightly for clarity.

BUG=137886
TEST=ash_unittests FramePainterTest.Basics


Review URL: https://chromiumcodereview.appspot.com/10808015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@147379 0039d316-1c4b-4281-b951-d872f2087c98