Make Chromoting client plugin always render at device DPI.
This addresses image quality issues when to a high-resolution device from a high-DPI but lower resolution client.  In this case Chromoting would down-scale to the plugin's DIP dimensions, only for Chrome to up-scale to device dimensions.

Follow-up CLs will add DPI information to the FrameConsumer interface, and try to optimize the plugin's up-scaling mode.

BUG=133349,135089


Review URL: https://chromiumcodereview.appspot.com/10785041

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@148012 0039d316-1c4b-4281-b951-d872f2087c98