aura: apply brightness filter last to improve perf
https://bugs.webkit.org/show_bug.cgi?id=92059 adds an optimization on color
matrix filters so that they can be combined in a single pass as long as the
resulting colors don't need to be clamped. Brightness needs clamping, so it
should go last.

This also fixes an uninitialized value that made us stick a grayscale filter randomly on layers.

BUG=138645
TEST=new startup animation


Review URL: https://chromiumcodereview.appspot.com/10833027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@148837 0039d316-1c4b-4281-b951-d872f2087c98