[Chromoting] Add wildcard to talkgadget in manifest permissions.
This grants permission to <anything>.talkgadget.google.com as well as allowing
simply talkgadget.google.com.

This is to prepare for (1) changing the default talkgadget name to something
specific to chromoting and (2) adding a policy setting that allows the
talkgadget name to be changed.

BUG=108448
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10789054

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@149105 0039d316-1c4b-4281-b951-d872f2087c98