Diagnostics UI: apply UI styles.
Add a navigation pane on the left (similar to the uber frame for history/extensions/settings pages) and apply the same style settings.

Note: the navigation pane is non-functional at the moment. For Diagnostics V1 (1st milestone), we will only have a single page for connectivity, which is always selected as the current page.

BUG=139442
TEST=lumpy


Review URL: https://chromiumcodereview.appspot.com/10824223

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@150793 0039d316-1c4b-4281-b951-d872f2087c98