gestures: Generate only either scroll-end or fling-start events at the end of a scroll gesture.
When a fling happens at the end of a scroll, instead of generating both a scroll-end
and a fling-start event, generate only fling-start event. Unfortunately, WebKit still
requires an explicit scroll-end gesture before a fling, so continue to do that as a
special case.

BUG=141460,141309

Review URL: https://chromiumcodereview.appspot.com/10826209

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@150852 0039d316-1c4b-4281-b951-d872f2087c98