gdata: Remove FindEntryByPathAndRunSync() from GDataWapiFeedLoader.
GDataWapiFeedLoader sould only focus on loading the feeds, without
doing the extra step to find an entry.

The number of callers of FindEntryByPathAndRunSync() is reduced from 4 to 2.

Along the way, make LoadDocumentFeedCallback parameter mandatory for
LoadFromServer() and friends. Looking at callers, it's safe to make this
mandatory.

BUG=141196, 126634
TEST=file manager works as before (loading from the server and from the cache)

Review URL: https://chromiumcodereview.appspot.com/10824303

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@151654 0039d316-1c4b-4281-b951-d872f2087c98