Regression test for anti-DDoS bugs in ChromeNetworkDelegate.
This ensures:
a) That when g_never_throttle_requests_ is false, only
extension requests may be throttled (rather than the inverse);
b) That setting the flag applies to all instances of
ChromeNetworkDelegate, not just the one it is set on.

If we'd had these tests earlier, they would have prevented the
bugs below from being introduced.

TBR=thakis@chromium.org
BUG=141081,141643


Review URL: https://chromiumcodereview.appspot.com/10828284

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@151683 0039d316-1c4b-4281-b951-d872f2087c98