Disable warnings in harfbuzz/ for Android when not using clang.
The Android NDK compilers, both on GCC 4.2 and on GCC 4.6, throw a warning about
incompatible pointer signedness in one of harfbuzz' APIs. That's where the fun
comes in: no flag is currently able to disable the warning when using Android's GCCs.
We therefore have to disable all warnings.

This is unfortunate, but we should still be getting sufficient coverage
for other warnings from Linux and Android-clang builds.

BUG=

Review URL: https://chromiumcodereview.appspot.com/10829354

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@151944 0039d316-1c4b-4281-b951-d872f2087c98