Cocoa: Implement media gallery dialog
This is the cocoa version of the media galleries permission dialog. The dialog is implemented as an NSAlert displayed as a constrained window. The gallery checkboxes are added inside a accessory view.

The views and GTK versions are here:
http://codereview.chromium.org/10828166/
https://chromiumcodereview.appspot.com/10826129

Screenshots: http://i.imgur.com/fedm0.png

BUG=134929
TBR=mihaip@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10834242

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@152128 0039d316-1c4b-4281-b951-d872f2087c98