Disable suggest field trial in tests.
See bug link. The test flakiness is because of being randomly selected into a
suggest field trial that causes the Google-returned suggestion to be ranked
above the what-you-typed entry. This CL attempts to disable the suggest field
trial by forcing a specific group (which at the moment doesn't have any
experimental behaviour).

BUG=143308
R=mpearson@chromium.org,nirnimesh@chromium.org
TEST=This test should no longer flake on the bots. If it does, then it's likely
that group #10 has been selected into a field trial. The fix would be to
disable the suggest field trials properly (as mentioned in the TODO).

Review URL: https://chromiumcodereview.appspot.com/10860022

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@152183 0039d316-1c4b-4281-b951-d872f2087c98