[Android] Implement vsync signal source
Implement a vsync signal source for locking compositor rendering
to the display frame period. It is based on the Choreographer API
introduced in Android Jellybean (API level 16).

Note that this patch does not tie the signal source to the compositor,
because all the necessary pieces have not been upstreamed yet.

BUG=http://crbug.com/137795
TEST=org.chromium.content.browser.test.VSyncMonitorTest


Review URL: https://chromiumcodereview.appspot.com/10854247

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@153214 0039d316-1c4b-4281-b951-d872f2087c98