Ash: Support "Next Song", "Previous Song", "Play/Pause" Multi Media buttons in our Media Player that is part of the File Manager right now.
Added shortcut handlers and plumbing to the Audio and Video player of the
Ash File Manager.

see also:
http://code.google.com/p/chromium/issues/detail?id=123739#c17

BUG=138745
TEST=Attach an (Wired USB) Apple keyboard or Windows Natural keyboard
to a Chromebook.
Either download three mp3 and an mp4 file to your chromebook or navigate
to a Google Drive folder that has these.
Bring up File Manager from the Launcher and click on an mp3.
Press the Previous Track, Play/Pause, Next Track keys on the MAC keyboard and
observe same behavior as clicking the corresponding Audio Control buttons
on the screen.
Click on an mp4 and press the Play/Pause key and observe same
behavior as clicking Play/Pause button on the screen.
Repeat for Windows Natural keyboard with the Play/Pause key. (Note
it does not have any prev./next track keys.)
PS: The Goldtouch USB keyboard works as well. The media keys require holding done the "Fn" key and F1, F2, or F3.

Review URL: https://chromiumcodereview.appspot.com/10823439

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@153502 0039d316-1c4b-4281-b951-d872f2087c98