Simplify handling of detached WebAccessibilityObjects.
Previously, every call to WebAccessibilityObject called an internal
updateBackingStore method, which could cause the WebAccessibilityObject
to be detached. It was inefficient to call it with every method call, and
also led to a heap-use-after-free.

With this patch, updateBackingStoreAndCheckValidity is called once before
accessing a WebAccessibilityObject, so we can next remove the internal
calls to updateBackingStore, fixing the heap-use-after-free.  This also
consolidates calls to isNull, isDetached, and isValid to just use
isDetached.

BUG=129301


Review URL: https://chromiumcodereview.appspot.com/10871085

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@153548 0039d316-1c4b-4281-b951-d872f2087c98