drive: Simplify DriveFileSystem::GetFileByResourceId()
This is a reland of crrev.com/153776, which was reverted due to crashes
in tests, which caused by misuse of scoped_ptr<>::Pass() again.

The number of callers of GetEntryByResourceIdAsync() in DriveFileSystem
is reduced from 2 to 1.

Along the way, make |get_file_callback| parameter mandatory for GetFileByPath()
GetFileByResourceId(), and related helper functions.

BUG=143518
TEST="make available offline" works as before, which is implemented on top of this function

Review URL: https://chromiumcodereview.appspot.com/10887017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@153928 0039d316-1c4b-4281-b951-d872f2087c98