Revert 154453 - Remove the translate pyauto test to chrome tests and all the supporting automation hooks. I've added new tests for the non-duplicate cases.
Most of the Chrome tests use a test harness to avoid network requests to the Translate server, which is inline with our testing phiolosophy (i.e. new bots don't even have network access). There is a browser_test in BrowserTest.PageLanguageDetection.

testTranslate is a duplicate of TranslateManagerTest.NormalTranslate
testNoTranslate is a duplicate of TranslateManagerTest.MultipleOnPageContents and TranslateManagerTest.DenyTranslateInPageNavigation
testNeverTranslateLanguage and testAlwaysTranslateLanguage are duplicates of TranslateManagerTest.BeforeTranslateExtraButtons
testNeverTranslateSite is a duplicate of TranslateManagerTest.NeverTranslateSitePref
testRevert is a duplicate of TranslateManagerTest.ServerReportsUnsupportedLanguage
testBarNotVisibleOnSSLErrorPage seems unnecessary, i.e. we don't check that infobar comes up for all of our internal pages
testBarNotVisibleOnEnglishPage is a duplicate of TranslateManagerTest.AutoTranslateOnNavigate
testTranslateDiffURLs, like in other pyauto test suites, seems unnecessary (i.e. we don't retest every feature across all schemes)
testNotranslateMetaTag is a duplicate of ChromeRenderViewTest.TranslatablePage
testToggleTranslateOption is a duplicate of TranslateManagerTest.AlwaysTranslateLanguagePref
testGoBackAndForwardToTranslatePage: the interesting bits are tested by ChromeRenderViewTest.BackToTranslatablePage
testForCrashedTab seems unnecessary, we don't verify that all infobars are gone when a page crashes
testTranslatePrefs is a duplicate of of TranslateManagerTest.TranslateEnabledPref
testAlwaysTranslateLanguageButton and testNeverTranslateLanguageButton are duplicates of TranslateManagerTest.BeforeTranslateExtraButtons
testChangeTargetLanguageAlwaysTranslate: I added a reload step to TranslateManagerTest.NormalTranslate to cover this
testHistoryNotTranslated and testDownloadsNotTranslated: I added a simple unittest for them
testSeveralLanguages is redundant with many of the above tests
testMultipleTabsAndWindows is not very useful
testAlwaysTranslateInIncognito is a duplicate of a subset of TranslateManagerTest.AlwaysTranslateLanguagePref
testNoTranslateInfobarAfterNeverTranslate is a duplicate of TranslateManagerTest.NeverTranslateLanguagePref
testGoBackAndForwardToTranslatePage is a duplicate of ChromeRenderViewTest.BackToTranslatablePage
testSessionRestore: I added a browser test for (BrowserTest.SessionRestore)

BUG=143637,95140,129802,12648,12639
Review URL: https://chromiumcodereview.appspot.com/10922002

TBR=jam@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10917042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@154464 0039d316-1c4b-4281-b951-d872f2087c98