[sync] Fix favicon add / load mechanism in integration tests
The sync test for favicon sync was flaky because of how we were adding a favicon directly via the BookmarkChangeProcessor. It turns out that we can now access the FaviconService for a profile via its factory, and that there is a much cleaner way of adding a favicon. There was also a problem with the way we were loading a favicon. The operations are asynchronous, as a result of which we must run the message loop until a non-empty favicon is loaded from the favicon service.

This patch updates the helper methods in bookmarks_helper.cc to directly set a favicon via FaviconService, and makes sure that a bookmark's favicon is fully (asynchronously) loaded before a favicon is returned.

BUG=94941
TEST=TwoClientBookmarksSyncTest.SC_AddFirstBMWithFavicon

Review URL: https://chromiumcodereview.appspot.com/10894043

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@154828 0039d316-1c4b-4281-b951-d872f2087c98